fix: openapi docs with correct endpoints
Added `/catalogs/:id/channel`, `/archives/:id/add-video` routes to catalog & archive respectively
