+++ src/app/api/archives/[archiveId]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest } from "next/server";
 
-import { getArchiveById, updateArchiveMeta } from "~/entities/archives";
+import { getArchiveById } from "~/entities/archives";
 import { NxResponse } from "~/shared/lib/next/nx-response";
 
 type ContextParams = {
@@ -18,15 +18,4 @@ export async function GET(_request: NextRequest, ctx: ContextParams) {
     data,
     200
   );
-}
-
-// TODO: Consider moving this to /archives/:id/update
-export async function PATCH(request: NextRequest, ctx: ContextParams) {
-  const { archiveId } = ctx.params;
-
-  const payload = await request.json();
-
-  const message = await updateArchiveMeta(archiveId, payload);
-
-  return NxResponse.success(message, {}, 201);
 }
