+++ src/app/api/archives/[archiveId]/update/route.ts
@@ -1,8 +1,6 @@
-import { revalidatePath } from "next/cache";
 import { NextRequest } from "next/server";
 
-import { addArchiveVideo } from "~/entities/archives";
-import { getUserIdCookie } from "~/shared/lib/next/get-cookie";
+import { updateArchiveMeta } from "~/entities/archives";
 import { NxResponse } from "~/shared/lib/next/nx-response";
 
 type ContextParams = {
@@ -11,17 +9,12 @@ type ContextParams = {
   };
 };
 
-// TODO: Consider changing this route to /archives/:id/add-video
 export async function PATCH(request: NextRequest, ctx: ContextParams) {
-  const userId = getUserIdCookie();
   const { archiveId } = ctx.params;
 
   const payload = await request.json();
-  const result = await addArchiveVideo(userId, archiveId, payload);
 
-  // Reset page cache when archives updates
-  revalidatePath("/explore/archives");
-  revalidatePath(`/a/${archiveId}`);
+  const message = await updateArchiveMeta(archiveId, payload);
 
-  return NxResponse.success(result, {}, 201);
+  return NxResponse.success(message, {}, 201);
 }
