+++ src/app/api/catalogs/[catalogId]/update/route.ts
@@ -1,12 +1,6 @@
-import { revalidatePath } from "next/cache";
 import { NextRequest } from "next/server";
 
-import {
-  deleteChannel,
-  getVideosByCatalog,
-  updateCatalogChannels,
-} from "~/entities/catalogs";
-import { getUserIdCookie } from "~/shared/lib/next/get-cookie";
+import { updateCatalogMeta } from "~/entities/catalogs";
 import { NxResponse } from "~/shared/lib/next/nx-response";
 
 type ContextParams = {
@@ -15,45 +9,12 @@ type ContextParams = {
   };
 };
 
-export async function DELETE(request: NextRequest, ctx: ContextParams) {
-  const userId = getUserIdCookie();
-  const { catalogId } = ctx.params;
-
-  const channels = await request.json();
-
-  await deleteChannel(userId, catalogId, channels);
-
-  revalidatePath(`/c/${catalogId}`);
-
-  return NxResponse.success<any>("Channel deleted successfully.", {}, 201);
-}
-
-// TODO: Consider moving this to /catalogs/:id/channel
 export async function PATCH(request: NextRequest, ctx: ContextParams) {
-  const userId = getUserIdCookie();
   const { catalogId } = ctx.params;
 
-  if (!catalogId) {
-    return NxResponse.fail(
-      "Catalog ID is missing from request params.",
-      {
-        code: "BAD_REQUEST",
-        details: "Catalog ID is missing from request params.",
-      },
-      400
-    );
-  }
+  const payload = await request.json();
 
-  const catalogPayload = await request.json();
+  const message = await updateCatalogMeta(catalogId, payload);
 
-  await updateCatalogChannels(userId, catalogId, catalogPayload);
-
-  // Update the catalog
-  getVideosByCatalog(catalogId);
-
-  // Revalidate the /explore route
-  revalidatePath("/explore/catalogs");
-  revalidatePath(`/c/${catalogId}`);
-
-  return NxResponse.success<any>("Channel list update successfully.", {}, 201);
+  return NxResponse.success(message, {}, 201);
 }
