+++ src/views/edit-archive/add-video-dialog.tsx
@@ -76,7 +76,7 @@ export default function AddVideoDialog({
         videoId: videoId,
       };
 
-      const resultAdd = await fetchApi(`/archives/${archiveId}/update`, {
+      const resultAdd = await fetchApi(`/archives/${archiveId}/add-video`, {
         method: "PATCH",
         body: JSON.stringify(videoMeta),
       });
